requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
pymongo
boto3
tensorflow
flask
pyyaml
python-dotenv